typemacros: check and reject too large arrays
Also correct the type of len from int to long.

+++ libubx/trig_utils.h
@@ -28,8 +28,8 @@ long cfg_getptr_trig_spec(const ubx_block_t *b,
 /* tstats port read/write helpers */
 long read_tstat(const ubx_port_t* p, struct ubx_tstat* val);
 int write_tstat(const ubx_port_t *p, const struct ubx_tstat *val);
-long read_tstat_array(const ubx_port_t* p, struct ubx_tstat* val, const int len);
-int write_tstat_array(const ubx_port_t* p, const struct ubx_tstat* val, const int len);
+long read_tstat_array(const ubx_port_t* p, struct ubx_tstat* val, const long len);
+int write_tstat_array(const ubx_port_t* p, const struct ubx_tstat* val, const long len);
 
 /**
  * struct trigger_info - basic trigger information
